// src/app/models/chamado.ts
export interface Chamado {
    id?:                   any;
    dataAbertura?:      string;
    dataFechamento?:    string;
    prioridade:         string;
    status:             string;
    titulo:             string;
    observacoes:        string;
    tecnico:               any;
    cliente:               any;
    nomeCliente:        string;
    nomeTecnico:        string;
}